feat(sheer-cycle): sheering added for 3/6 directions: xy:+x xz:+x yz:+y
