+++ bin/shear-cycle
@@ -1,11 +1,11 @@
 #!/usr/bin/env bash
 set -euo pipefail
 [ -z ${8+x} ] && (
-    echo "USAGE: <slope+xy:double> <slope+xz:double> <slope+yz:double>"
+    echo "USAGE: <shear+xy:double> <shear+xz:double> <shear+yz:double>"
     echo "       <cycle:int(0|1)> <antialias:int(0|1)>"
     echo "       <source_directory:str> <target_directory:str>"
     echo "       <(JPEG|PNG|TIFF)>"
     false)
 /usr/bin/env mvn -e -q exec:java \
-    -D'exec.mainClass'='ca.petrographic.mains.SlopeCycle' \
+    -D'exec.mainClass'='ca.petrographic.mains.ShearCycle' \
     -D'exec.args'="$1 $2 $3 $4 $5 $6 $7 $8"
